// src/types/resources/Provenance.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TReference } from '../partials/Reference';
import { TPeriod } from '../partials/Period';
import { TDateTime } from '../simpleTypes/DateTime';
import { TInstant } from '../simpleTypes/Instant';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TProvenanceAgent } from '../partials/ProvenanceAgent';
import { TProvenanceEntity } from '../partials/ProvenanceEntity';
import { TSignature } from '../partials/Signature';

export type TProvenance = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  target: TReference[];
  occurredPeriod?: TPeriod;
  occurredDateTime?: TDateTime;
  recorded: TInstant;
  policy?: TUri[];
  location?: TReference;
  reason?: TCodeableConcept[];
  activity?: TCodeableConcept;
  agent: TProvenanceAgent[];
  entity?: TProvenanceEntity[];
  signature?: TSignature[];
};